+++ app/models/gamesModel.js
@@ -2,6 +2,11 @@ const mongoose = require("mongoose");
 
 const gamesSchema = new mongoose.Schema(
     {
+        username: {
+            type: String,
+            unique: true
+        },
+        name: String,
         userMove: String,
         computerMove: String,
         result: String,
