+++ app/routes/mainRoutes.js
@@ -1,11 +1,14 @@
 const express = require('express');
 const router = express.Router();
-const mainController = require('../controllers/mainControllers');
+const authController = require('../controllers/authController');
+const userController = require('../controllers/userController');
+const gameController = require('../controllers/gameController');
 
-router.post('/signup', mainController.signup);
-router.post('/login', mainController.login);
-router.post('/game/:username', mainController.game);
-router.delete('/logout/:username', mainController.logout);
+router.post('/signup', userController.signup);
+router.post('/login', authController.login);
+router.post('/game/:username', gameController.game);
+router.post('/game/reset/:username', gameController.reset);
+router.delete('/logout/:username', authController.logout);
 
 
 module.exports = router;
